The main plugin code for streaming steering data from GP-Bikes. Modified from PiBoSo's example plugin.
This is the main plugin which should be compiled as a dll with extension .dlo and added to the plugins folder of GP-Bikes. Since this is early stages there's a lot of junk code I've left in for debugging purposes but I'm sure it's pretty easy to follow since PiBoSo has commented it so well.